AutoUpgrade should work
Runs before shutdown and not reboot
activates on Fridays at 20:00 and on next shutdown will upgrade. If the PC is off, it will be activated in the next day.

Problem: If instead of a poweroff, a reboot is issued, the upgrade won't happen and the timer won't be activated after the reboot, having to wait for next week to hopefully get the upgrade